- Move util into pose_tensorflow package to be more similar to head   repo
- Fix demo/singleperson.py by making initialization in
  pose_tensorflow/__init_.py explicit instead of automatically on
  import

- Put back demo/pose_cfg.yaml to be similar to head repo
