+++ demo/singleperson.py
@@ -5,10 +5,10 @@ sys.path.append(os.path.dirname(__file__) + "/../")
 
 from scipy.misc import imread
 
-from util.config import load_config
-from nnet import predict
-from util import visualize
-from dataset.pose_dataset import data_to_input
+from pose_tensorflow.util.config import load_config
+from pose_tensorflow.nnet import predict
+from pose_tensorflow.util import visualize
+from pose_tensorflow.dataset.pose_dataset import data_to_input
 
 
 cfg = load_config("demo/pose_cfg.yaml")
