+++ pose_tensorflow/util/default_config.py
