+++ pose_tensorflow/util/mscoco_util.py
