+++ pose_tensorflow/util/visualize.py
